feat: move modal contents to src/components/modalContent
