+++ src/pages/listPage/ListPage.jsx
@@ -4,12 +4,12 @@ import Header from '@/components/Header';
 import CreditSection from '@/pages/listPage/credit/CreditSection';
 import { getCredits } from '@/utils/creditStorage';
 import RechargeModalContent from '@/components/modalContent/RechargeModalContent';
-import CreditRechargeSuccess from '@/pages/listPage/credit/CreditRechargeSuccess';
+import CreditRechargeSuccess from '@/components/modalContent/CreditRechargeSuccess';
 import CreditShortageModalContent from '@/components/modalContent/CreditShortageModalContent';
 import DonationsList from '@/pages/listPage/donation/DonationsList';
 import DonationModalContent from '@/components/modalContent/DonationModalContent';
 import leftTopGradient from '@/assets/images/leftTopGradient.png';
-import DonationSuccess from '@/pages/listPage/donation/DonationSuccess';
+import DonationSuccess from '@/components/modalContent/DonationSuccess';
 import MonthlyChartSection from '@/pages/listPage/monthlyChart/MonthlyChartSection';
 
 function ListPage() {
